Move TempRValueOpt from CopyForwarding.cpp -> TempRValueElimination.cpp.
The pass is standalone from CopyForwarding and is big enough to stand on its
own.

This changes is NFC beyond clang-format + xcode rename.
