+++ lib/SILOptimizer/Transforms/CopyForwarding.cpp
@@ -1538,403 +1538,12 @@ class CopyForwardingPass : public SILFunctionTransform
       invalidateAnalysis(SILAnalysis::InvalidationKind::FunctionBody);
     } else {
       invalidateAnalysis(SILAnalysis::InvalidationKind::CallsAndInstructions);
-      }
+    }
   }
 };
-
-/// Temporary RValue Optimization
-///
-/// Peephole optimization to eliminate short-lived immutable temporary copies.
-/// This handles a common pattern generated by SILGen where temporary RValues
-/// are emitted as copies...
-///
-///   %temp = alloc_stack $T
-///   copy_addr %src to [initialization] %temp : $*T
-///   // no writes to %src and %temp
-///   destroy_addr %temp : $*T
-///   dealloc_stack %temp : $*T
-///
-/// This differs from the copy forwarding algorithm because it handles
-/// copy source and dest lifetimes that are unavoidably overlappying. Instead,
-/// it finds cases in which it is easy to determine that the source is
-/// unmodified during the copy destination's lifetime. Thus, the destination can
-/// be viewed as a short-lived "rvalue".
-class TempRValueOptPass : public SILFunctionTransform {
-  AliasAnalysis *AA = nullptr;
-
-  bool collectLoads(Operand *UserOp, SILInstruction *UserInst,
-                           SingleValueInstruction *Addr,
-                           SILValue srcObject,
-                           llvm::SmallPtrSetImpl<SILInstruction *> &LoadInsts);
-
-  bool checkNoSourceModification(CopyAddrInst *copyInst,
-                       const llvm::SmallPtrSetImpl<SILInstruction *> &useInsts);
-
-  bool checkTempObjectDestroy(AllocStackInst *tempObj, CopyAddrInst *copyInst);
-
-  bool tryOptimizeCopyIntoTemp(CopyAddrInst *copyInst);
-
-  void run() override;
-};
-
-/// The main entry point of the pass.
-void TempRValueOptPass::run() {
-  if (getFunction()->hasOwnership())
-    return;
-
-  LLVM_DEBUG(llvm::dbgs() << "Copy Peephole in Func "
-                          << getFunction()->getName() << "\n");
-
-  AA = PM->getAnalysis<AliasAnalysis>();
-  bool Changed = false;
-
-  // Find all copy_addr instructions.
-  for (auto &BB : *getFunction()) {
-    auto II = BB.begin();
-    while (II != BB.end()) {
-      auto *CopyInst = dyn_cast<CopyAddrInst>(&*II);
-
-      if (CopyInst) {
-        // In case of success, this may delete instructions, but not the
-        // CopyInst itself.
-        Changed |= tryOptimizeCopyIntoTemp(CopyInst);
-      }
-
-      // Increment the instruction iterator here. We can't do it at the begin of
-      // the loop because the instruction after CopyInst might be deleted in
-      // in tryOptimizeCopyIntoTemp. We can't do it at the end of the loop
-      // because the CopyInst might be deleted in the following code.
-      ++II;
-
-      // Remove identity copies which are a result of this optimization.
-      if (CopyInst && CopyInst->getSrc() == CopyInst->getDest()) {
-        // This is either the CopyInst which just got optimized or it is a
-        // follow-up from an earlier iteration, where another copy_addr copied
-        // the temporary back to the source location.
-        CopyInst->eraseFromParent();
-      }
-    }
-  }
-
-  if (Changed) {
-    invalidateAnalysis(SILAnalysis::InvalidationKind::Instructions);
-  }
-}
-
-/// Transitively explore all data flow uses of the given \p address until
-/// reaching a load or returning false.
-///
-/// Any user opcode recognized by collectLoads must be replaced correctly later
-/// during tryOptimizeCopyIntoTemp. If it is possible for any use to destroy the
-/// value in \p address, then that use must be removed or made non-destructive
-/// after the copy is removed and its operand is replaced.
-///
-/// Warning: To preserve the original object lifetime, tryOptimizeCopyIntoTemp
-/// must assume that there are no holes in lifetime of the temporary stack
-/// location at \address. The temporary must be initialized by the original copy
-/// and never written to again. Therefore, collectLoads disallows any operation
-/// that may write to memory at \p address.
-bool TempRValueOptPass::collectLoads(
-    Operand *userOp, SILInstruction *user, SingleValueInstruction *address,
-    SILValue srcObject,
-    llvm::SmallPtrSetImpl<SILInstruction *> &loadInsts) {
-  // All normal uses (loads) must be in the initialization block.
-  // (The destroy and dealloc are commonly in a different block though.)
-  if (user->getParent() != address->getParent())
-    return false;
-
-  // Only allow uses that cannot destroy their operand. We need to be sure
-  // that replacing all this temporary's uses with the copy source doesn't
-  // destroy the source. This way, we know that the destroy_addr instructions
-  // that we recorded cover all the temporary's lifetime termination points.
-  //
-  // Currently this includes address projections, loads, and in_guaranteed uses
-  // by an apply.
-  //
-  // TODO: handle non-destructive projections of enums
-  // (unchecked_take_enum_data_addr of Optional is nondestructive.)
-  switch (user->getKind()) {
-  default:
-    LLVM_DEBUG(llvm::dbgs() << "  Temp use may write/destroy its source"
-                            << *user);
-    return false;
-
-  case SILInstructionKind::ApplyInst:
-  case SILInstructionKind::TryApplyInst: {
-    ApplySite apply(user);
-
-    // Check if the function can just read from userOp.
-    auto Convention = apply.getArgumentConvention(*userOp);
-    if (!Convention.isGuaranteedConvention()) {
-      LLVM_DEBUG(llvm::dbgs() << "  Temp consuming use may write/destroy "
-                 "its source" << *user);
-      return false;
-    }
-
-    // Check if there is another function argument, which is inout which might
-    // modify the source of the copy_addr.
-    //
-    // When a use of the temporary is an apply, then we need to prove that the
-    // function called by the apply cannot modify the temporary's source
-    // value. By design, this should be handled by
-    // `checkNoSourceModification`. However, this would be too conservative
-    // since it's common for the apply to have an @out argument, and alias
-    // analysis cannot prove that the @out does not alias with `src`. Instead,
-    // `checkNoSourceModification` always avoids analyzing the current use, so
-    // applies need to be handled here. We already know that an @out cannot
-    // alias with `src` because the `src` value must be initialized at the point
-    // of the call. Hence, it is sufficient to check specifically for another
-    // @inout that might alias with `src`.
-    auto calleeConv = apply.getSubstCalleeConv();
-    unsigned calleeArgIdx = apply.getCalleeArgIndexOfFirstAppliedArg();
-    for (Operand &operand : apply.getArgumentOperands()) {
-      auto argConv = calleeConv.getSILArgumentConvention(calleeArgIdx);
-      if (argConv.isInoutConvention()) {
-        if (!AA->isNoAlias(operand.get(), srcObject)) {
-          return false;
-        }
-      }
-      ++calleeArgIdx;
-    }
-
-    // Everything is okay with the function call. Register it as a "load".
-    loadInsts.insert(user);
-    return true;
-  }
-  case SILInstructionKind::OpenExistentialAddrInst: {
-    // We only support open existential addr if the access is immutable.
-    auto *oeai = cast<OpenExistentialAddrInst>(user);
-    if (oeai->getAccessKind() != OpenedExistentialAccess::Immutable) {
-      LLVM_DEBUG(llvm::dbgs() << "  Temp consuming use may write/destroy "
-                 "its source" << *user);
-      return false;
-    }
-    LLVM_FALLTHROUGH;
-  }
-  case SILInstructionKind::StructElementAddrInst:
-  case SILInstructionKind::TupleElementAddrInst: {
-    // Transitively look through projections on stack addresses.
-    auto proj = cast<SingleValueInstruction>(user);
-    for (auto *projUseOper : proj->getUses()) {
-      auto *user = projUseOper->getUser();
-      if (user->isTypeDependentOperand(*projUseOper))
-        continue;
-
-      if (!collectLoads(projUseOper, user, proj, srcObject,
-                        loadInsts))
-        return false;
-    }
-    return true;
-  }
-
-  case SILInstructionKind::LoadInst:
-  case SILInstructionKind::LoadBorrowInst:
-    // Loads are the end of the data flow chain. The users of the load can't
-    // access the temporary storage.
-    loadInsts.insert(user);
-    return true;
-
-  case SILInstructionKind::CopyAddrInst: {
-    // copy_addr which read from the temporary are like loads.
-    auto *copyFromTmp = cast<CopyAddrInst>(user);
-    if (copyFromTmp->getDest() == address) {
-      LLVM_DEBUG(llvm::dbgs() << "  Temp written or taken" << *user);
-      return false;
-    }
-    loadInsts.insert(copyFromTmp);
-    return true;
-  }
-  }
-}
-
-/// Checks if the copy's source can be modified within the temporary's lifetime.
-///
-/// Unfortunately, we cannot simply use the destroy points as the lifetime end,
-/// because they can be in a different basic block (that's what SILGen
-/// generates). Instead we guarantee that all normal uses are within the block
-/// of the temporary and look for the last use, which effectively ends the
-/// lifetime.
-bool TempRValueOptPass::checkNoSourceModification(CopyAddrInst *copyInst,
-                     const llvm::SmallPtrSetImpl<SILInstruction *> &useInsts) {
-  unsigned NumLoadsFound = 0;
-  auto iter = std::next(copyInst->getIterator());
-  // We already checked that the useful lifetime of the temporary ends in
-  // the initialization block.
-  auto iterEnd = copyInst->getParent()->end();
-  for (; iter != iterEnd; ++iter) {
-    SILInstruction *I = &*iter;
-
-    if (useInsts.count(I))
-      NumLoadsFound++;
-
-    // If this is the last use of the temp we are ok. After this point,
-    // modifications to the source don't matter anymore.
-    if (NumLoadsFound == useInsts.size())
-      return true;
-
-    if (AA->mayWriteToMemory(I, copyInst->getSrc())) {
-      LLVM_DEBUG(llvm::dbgs() << "  Source modified by" << *iter);
-      return false;
-    }
-  }
-  // For some reason, not all normal uses have been seen between the copy and
-  // the end of the initialization block. We should never reach here.
-  return false;
-}
-
-/// Return true if the \p tempObj, which is initialized by \p copyInst, is
-/// destroyed in an orthodox way.
-///
-/// When tryOptimizeCopyIntoTemp replaces all of tempObj's uses, it assumes that
-/// the object is initialized by the original copy and directly destroyed on all
-/// paths by one of the recognized 'destroy_addr' or 'copy_addr [take]'
-/// operations. This assumption must be checked. For example, in non-OSSA,
-/// it is legal to destroy an in-memory object by loading the value and
-/// releasing it. Rather than detecting unbalanced load releases, simply check
-/// that tempObj is destroyed directly on all paths.
-bool TempRValueOptPass::checkTempObjectDestroy(AllocStackInst *tempObj,
-                                               CopyAddrInst *copyInst) {
-  // If the original copy was a take, then replacing all uses cannot affect
-  // the lifetime.
-  if (copyInst->isTakeOfSrc())
-    return true;
-
-  // ValueLifetimeAnalysis is not normally used for address types. It does not
-  // reason about the lifetime of the in-memory object. However the utility can
-  // be abused here to check that the address is directly destroyed on all
-  // paths. collectLoads has already guaranteed that tempObj's lifetime has no
-  // holes/reinitializations.
-  SmallVector<SILInstruction *, 8> users;
-  for (auto result : tempObj->getResults()) {
-    for (Operand *operand : result->getUses()) {
-      SILInstruction *user = operand->getUser();
-      if (user == copyInst)
-        continue;
-      if (isa<DeallocStackInst>(user))
-        continue;
-      users.push_back(user);
-    }
-  }
-  // Find the boundary of tempObj's address lifetime, starting at copyInst.
-  ValueLifetimeAnalysis vla(copyInst, users);
-  ValueLifetimeAnalysis::Frontier tempAddressFrontier;
-  if (!vla.computeFrontier(tempAddressFrontier,
-                           ValueLifetimeAnalysis::DontModifyCFG)) {
-    return false;
-  }
-  // Check that the lifetime boundary ends at direct destroy points.
-  for (SILInstruction *frontierInst : tempAddressFrontier) {
-    auto pos = frontierInst->getIterator();
-    // If the frontier is at the head of a block, then either it is an
-    // unexpected lifetime exit, or the lifetime ended at a
-    // terminator. TempRValueOptPass does not handle either case.
-    if (pos == frontierInst->getParent()->begin())
-      return false;
-
-    // Look for a known destroy point as described in the funciton level
-    // comment. This whitelist can be expanded as more cases are handled in
-    // tryOptimizeCopyIntoTemp during copy replacement.
-    SILInstruction *lastUser = &*std::prev(pos);
-    if (isa<DestroyAddrInst>(lastUser))
-      continue;
-
-    if (auto *cai = dyn_cast<CopyAddrInst>(lastUser)) {
-      assert(cai->getSrc() == tempObj && "collectLoads checks for writes");
-      assert(!copyInst->isTakeOfSrc() && "checked above");
-      if (cai->isTakeOfSrc())
-        continue;
-    }
-    return false;
-  }
-  return true;
-}
-
-/// Tries to perform the temporary rvalue copy elimination for \p copyInst
-bool TempRValueOptPass::tryOptimizeCopyIntoTemp(CopyAddrInst *copyInst) {
-  if (!copyInst->isInitializationOfDest())
-    return false;
-
-  auto *tempObj = dyn_cast<AllocStackInst>(copyInst->getDest());
-  if (!tempObj)
-    return false;
-
-  assert(tempObj != copyInst->getSrc() &&
-           "can't initialize temporary with itself");
-
-  // Scan all uses of the temporary storage (tempObj) to verify they all refer
-  // to the value initialized by this copy. It is sufficient to check that the
-  // only users that modify memory are the copy_addr [initialization] and
-  // destroy_addr.
-  llvm::SmallPtrSet<SILInstruction *, 8> loadInsts;
-  for (auto *useOper : tempObj->getUses()) {
-    SILInstruction *user = useOper->getUser();
-
-    if (user == copyInst)
-      continue;
-
-    // Destroys and deallocations are allowed to be in a different block.
-    if (isa<DestroyAddrInst>(user) || isa<DeallocStackInst>(user))
-      continue;
-
-    if (!collectLoads(useOper, user, tempObj, copyInst->getSrc(), loadInsts))
-      return false;
-  }
-
-  // Check if the source is modified within the lifetime of the temporary.
-  if (!checkNoSourceModification(copyInst, loadInsts))
-    return false;
-
-  if (!checkTempObjectDestroy(tempObj, copyInst))
-    return false;
-
-  LLVM_DEBUG(llvm::dbgs() << "  Success: replace temp" << *tempObj);
-
-  // Do a "replaceAllUses" by either deleting the users or replacing them with
-  // the source address. Note: we must not delete the original copyInst because
-  // it would crash the instruction iteration in run(). Instead the copyInst
-  // gets identical Src and Dest operands.
-  while (!tempObj->use_empty()) {
-    Operand *use = *tempObj->use_begin();
-    SILInstruction *user = use->getUser();
-    switch (user->getKind()) {
-    case SILInstructionKind::DestroyAddrInst:
-      if (copyInst->isTakeOfSrc()) {
-        use->set(copyInst->getSrc());
-      } else {
-        user->eraseFromParent();
-      }
-      break;
-    case SILInstructionKind::DeallocStackInst:
-      user->eraseFromParent();
-      break;
-    case SILInstructionKind::CopyAddrInst: {
-      auto *CAI = cast<CopyAddrInst>(user);
-      if (CAI != copyInst) {
-        assert(CAI->getSrc() == tempObj);
-        if (CAI->isTakeOfSrc() && !copyInst->isTakeOfSrc())
-          CAI->setIsTakeOfSrc(IsNotTake);
-      }
-      use->set(copyInst->getSrc());
-      break;
-    }
-    // ASSUMPTION: no operations that may be handled by this default clause can
-    // destroy tempObj. This includes operations that load the value from memory
-    // and release it.
-    default:
-      use->set(copyInst->getSrc());
-      break;
-    }
-  }
-  tempObj->eraseFromParent();
-  return true;
-}
 
 } // end anonymous namespace
 
 SILTransform *swift::createCopyForwarding() {
   return new CopyForwardingPass();
-}
-
-SILTransform *swift::createTempRValueOpt() {
-  return new TempRValueOptPass();
 }
